- Refactor project to es6 module. - Add tests
