+++ src/request/request.js
@@ -1,9 +1,8 @@
-const fetch = require('node-fetch');
-
-const requestSource = async (url) => {
+export async function request(url) {
     const response = await fetch(url);
     if (!response.ok) {
-        throw { code: response.status, message: response.statusText, url: url }
+        // throw { code: response.status, message: response.statusText, url: url }
+        throw new Error(response.statusText);
     }
     const result = await response.text();
     return {
@@ -12,5 +11,3 @@ const requestSource = async (url) => {
         body: result
     }
 }
-
-module.exports = { requestSource }
